requirements.txt
python-dotenv==0.21.0
pandas==1.1.5
openai